src/main/java/com/cite/employee_managment/service/AddressServiceImpl.java
package com.cite.employee_managment.service;

import com.cite.employee_managment.model.Address;
import com.cite.employee_managment.repo.AddressRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class AddressServiceImpl implements AddressService {

    @Autowired
    private AddressRepository addressRepository;

    @Override
    public Address save(Address address) {
        return addressRepository.save(address);
    }
}
